todo arreglado

+++ ProyectoTaw/src/main/java/es/uma/proyectotaw/controller/RutinaEntrenadorController.java
@@ -10,8 +10,11 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import java.util.ArrayList;
 import java.util.List;
 
+// Pablo Pardo Fernández - 80% (Listar/ Crear/Borrar/Ver/Guardar rutinas)
+//Alba Ruiz Gutiérrez
 @Controller
 public class RutinaEntrenadorController extends BaseController{
 
@@ -29,6 +32,15 @@ public class RutinaEntrenadorController extends BaseController{
 
     @Autowired
     private RutinaAsignadaRepository rutinaAsignadaRepository;
+
+    @Autowired
+    private SesionentrenamientoHasSesionejercicioRepository sesionentrenamientoHasSesionejercicioRepository;
+
+    @Autowired
+    private ValoracionRepository valoracionRepository;
+
+    @Autowired
+    private SesionejercicioRepository sesionejercicioRepository;
 
     @GetMapping("/entrenadorMain")
     public String doEntrenadorMain(Model model, HttpSession session) {
@@ -52,6 +64,30 @@ public class RutinaEntrenadorController extends BaseController{
         return "rutinasEntrenador";
     }
 
+    @PostMapping("/entrenadorMain/rutinas/filtrar")
+    public String doFiltrar(@RequestParam("filtro") String filtro, Model model, HttpSession session) {
+        if (!estaAutenticado(session)) return "redirect:/acceso";
+        UsuarioEntity usuario = (UsuarioEntity) session.getAttribute("usuario");
+        List<RutinaPredefinidaEntity> listaRutinas = rutinaPredefinidaRepository.findByUsuario(usuario);
+        List<RutinaPredefinidaEntity> listaFiltrada = new ArrayList<>();
+
+        for (RutinaPredefinidaEntity rutina : listaRutinas) {
+            if (rutina.getNombre() != null && rutina.getNombre().toLowerCase().contains(filtro.toLowerCase()) ||
+                    rutina.getObjetivos() != null && rutina.getObjetivos().toLowerCase().contains(filtro.toLowerCase())){
+                listaFiltrada.add(rutina);
+            }
+        }
+
+        if(filtro.isEmpty()){
+            listaFiltrada = listaRutinas;
+        }
+
+        model.addAttribute("filtro", filtro);
+        model.addAttribute("rutinas", listaFiltrada);
+        model.addAttribute("entrenador", usuario);
+        return "rutinasEntrenador";
+    }
+
     @GetMapping("/entrenadorMain/rutinas/crear")
     public String doCrearRutina(Model model, HttpSession session) {
         if(!estaAutenticado(session)) return "redirect:/acceso";
@@ -70,15 +106,32 @@ public class RutinaEntrenadorController extends BaseController{
     @GetMapping("/entrenadorMain/rutinas/borrar")
     public String doBorrarRutina(Model model, HttpSession session, @RequestParam("id") Integer id) {
         if(!estaAutenticado(session)) return "redirect:/acceso";
-        RutinaPredefinidaEntity rutina = rutinaPredefinidaRepository.findById(id).get();
+        RutinaPredefinidaEntity rutina = rutinaPredefinidaRepository.findById(id).orElse(null);
+        if (rutina == null) return "redirect:/entrenadorMain/rutinas"; // Rutina no encontrada
+
+        // Obtener las sesiones de entrenamiento asociadas a la rutina
         List<RutinaSesionentrenamientoEntity> rutinasSesiones = rutinaSesionentrenamientoRepository.findByRutinaPredefinidaOrderByPosicion(rutina);
+
+        // Obtener las rutinas asignadas asociadas a la rutina
+        List<RutinaAsignadaEntity> rutinasAsignadas = rutinaAsignadaRepository.findByRutinaPredefinida(rutina);
+
         for (RutinaSesionentrenamientoEntity rutinaSesion : rutinasSesiones) {
+            // Eliminar la relación entre la rutina y la sesión de entrenamiento
             rutinaSesionentrenamientoRepository.delete(rutinaSesion);
         }
-        List<RutinaAsignadaEntity> rutinasAsignadas = rutinaAsignadaRepository.findByRutinaPredefinida(rutina);
+
+        // Eliminar las rutinas asignadas
         for (RutinaAsignadaEntity rutinaAsignada : rutinasAsignadas) {
+            // Eliminar todas las valoraciones asociadas a la rutina asignada
+            List<ValoracionEntity> valoraciones = valoracionRepository.findByRutinaAsignada(rutinaAsignada);
+            for (ValoracionEntity valoracion : valoraciones) {
+                valoracionRepository.delete(valoracion);
+            }
+            // Finalmente eliminar la rutina asignada
             rutinaAsignadaRepository.delete(rutinaAsignada);
         }
+
+        // Finalmente, eliminar la rutina
         rutinaPredefinidaRepository.delete(rutina);
         return "redirect:/entrenadorMain/rutinas";
     }
